backend/pyproject.toml
[tool.poetry]
name = "GodsEye"
version = "0.1.0"
description = ""
authors = ["Subham Singh <geekysubham@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"
requests = "^2.32.3"
psutil = "^6.1.1"
python-dateutil = "^2.9.0.post0"
pydantic-extra-types = "^2.5.0"
email-validator = "^2.1.0"
watchdog = "^6.0.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.9.1"
pre-commit = "^4.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
